# Conflicts:
#	.github/workflows/gradle.yml
#	AUTHORS
#	README.md
#	build.gradle
#	core/build.gradle
#	core/src/main/java/org/bitcoinj/params/MainNetParams.java
#	core/src/main/java/org/litecoinj/core/AddressV1Message.java
#	core/src/main/java/org/litecoinj/core/AddressV2Message.java
#	core/src/main/java/org/litecoinj/core/Block.java
#	core/src/main/java/org/litecoinj/core/FeeFilterMessage.java
#	core/src/main/java/org/litecoinj/core/FullPrunedBlockChain.java
#	core/src/main/java/org/litecoinj/core/NetworkParameters.java
#	core/src/main/java/org/litecoinj/core/SendAddrV2Message.java
#	core/src/main/java/org/litecoinj/core/VersionMessage.java
#	core/src/main/java/org/litecoinj/params/RegTestParams.java
#	core/src/main/java/org/litecoinj/params/TestNet3Params.java
#	core/src/main/java/org/litecoinj/params/UnitTestParams.java
#	core/src/main/java/org/litecoinj/script/Script.java
#	core/src/main/java/org/litecoinj/utils/MonetaryFormat.java
#	core/src/main/java/org/litecoinj/wallet/Protos.java
#	core/src/main/java/org/litecoinj/wallet/WalletProtobufSerializer.java
#	core/src/main/resources/org.bitcoin.production.checkpoints.txt
#	core/src/main/resources/org.bitcoin.test.checkpoints.txt
#	core/src/test/java/org/bitcoinj/core/VersionMessageTest.java
#	core/src/test/java/org/bitcoinj/params/AbstractBitcoinNetParamsTest.java
#	core/src/test/java/org/bitcoinj/utils/MonetaryFormatTest.java
#	core/src/test/java/org/bitcoinj/wallet/WalletTest.java
#	examples/src/main/java/org/bitcoinj/examples/PeerMonitor.java
#	tools/src/main/java/org/bitcoinj/tools/WalletTool.java
#	wallettemplate/src/main/java/org/bitcoinj/walletfx/utils/BitcoinUIModel.java
#	wallettemplate/src/main/java/wallettemplate/Main.java
#	wallettemplate/src/main/java/wallettemplate/MainController.java
#	wallettemplate/src/main/java/wallettemplate/SendMoneyController.java
#	wallettemplate/src/main/java/wallettemplate/WalletSetPasswordController.java
#	wallettemplate/src/main/java/wallettemplate/WalletSettingsController.java

+++ core/src/main/java/org/litecoinj/core/HeadersMessage.java
@@ -66,7 +66,7 @@ public class HeadersMessage extends Message {
 
     @Override
     protected void parse() throws ProtocolException {
-        long numHeaders = readVarInt();
+        int numHeaders = readVarInt().intValue();
         if (numHeaders > MAX_HEADERS)
             throw new ProtocolException("Too many headers: got " + numHeaders + " which is larger than " +
                                          MAX_HEADERS);
